Refactored UserIdentity as a base class which is inherited by HashedDii/FirstLevelHash/RawUidIdentity sub-classes
