+++ src/main/java/com/uid2/operator/model/AdvertisingTokenInput.java
@@ -1,6 +1,8 @@
 package com.uid2.operator.model;
 
 import java.time.Instant;
+
+import com.uid2.operator.model.userIdentity.RawUidIdentity;
 import com.uid2.shared.model.TokenVersion;
 
 public class AdvertisingTokenInput extends VersionedToken {
