+++ src/main/java/com/uid2/operator/model/RefreshTokenInput.java
@@ -1,6 +1,8 @@
 package com.uid2.operator.model;
 
 import java.time.Instant;
+
+import com.uid2.operator.model.userIdentity.FirstLevelHashIdentity;
 import com.uid2.shared.model.TokenVersion;
 
 public class RefreshTokenInput extends VersionedToken {
