+++ src/main/java/com/uid2/operator/model/userIdentity/RawUidIdentity.java
@@ -1,16 +1,14 @@
-package com.uid2.operator.model;
+package com.uid2.operator.model.userIdentity;
+
+import com.uid2.operator.model.IdentityScope;
+import com.uid2.operator.model.IdentityType;
 
 import java.time.Instant;
 import java.util.Arrays;
 
 // A raw UID is stored inside
-public class RawUidIdentity implements UserIdentity {
-    public final IdentityScope identityScope;
-    public final IdentityType identityType;
+public class RawUidIdentity extends UserIdentity {
     public final byte[] rawUid;
-    public final int privacyBits;
-    public final Instant establishedAt;
-    public final Instant refreshedAt;
 
     public RawUidIdentity(IdentityScope identityScope, IdentityType identityType, byte[] rawUid, int privacyBits,
                           Instant establishedAt, Instant refreshedAt) {
@@ -27,9 +25,4 @@ public class RawUidIdentity implements UserIdentity {
                 this.identityType.equals(that.identityType) &&
                 Arrays.equals(this.rawUid, that.rawUid);
     }
-
-    public IdentityScope GetIdentityScope() { return identityScope; }
-    public IdentityType GetIdentityType() { return identityType; }
-    public Instant GetEstablishedAt() { return establishedAt; };
-    public Instant GetIRefreshedAt() { return refreshedAt; }
 }
