+++ src/main/java/com/uid2/operator/store/CloudSyncOptOutStore.java
@@ -4,8 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.uid2.operator.Const;
-import com.uid2.operator.model.FirstLevelHashIdentity;
-import com.uid2.operator.model.HashedDiiIdentity;
+import com.uid2.operator.model.userIdentity.FirstLevelHashIdentity;
 import com.uid2.operator.service.EncodingUtils;
 import com.uid2.shared.Utils;
 import com.uid2.shared.cloud.CloudStorageException;
