+++ src/test/java/com/uid2/operator/benchmark/TokenEndecBenchmark.java
@@ -1,6 +1,7 @@
 package com.uid2.operator.benchmark;
 
 import com.uid2.operator.model.*;
+import com.uid2.operator.model.userIdentity.HashedDiiIdentity;
 import com.uid2.operator.service.EncryptedTokenEncoder;
 import com.uid2.operator.service.IUIDOperatorService;
 import org.openjdk.jmh.annotations.Benchmark;
